third_party/WebKit/LayoutTests/external/wpt/html/semantics/embedded-content/the-img-element/img.complete-expected.txt

This is a testharness.js-based test.
PASS img src omitted
PASS img src empty
FAIL async src complete test assert_equals: expected "/3.jpg" but got "/http://web-platform.test:8001/html/semantics/embedded-content/the-img-element/3.jpg"
FAIL IDL attribute complete returns true when image resource has been fetched but not run yet & image is not in broken state assert_unreached: .complete didn't change to true Reached unreachable code
PASS async src broken test
Harness: the test ran to completion.
